Move reds tests from Player.spec.ts to Turmoi.spec.ts.
