+++ tests/turmoil/Turmoil.spec.ts
@@ -12,6 +12,7 @@ import {TestingUtils} from '../TestingUtils';
 import {TestPlayers} from '../TestPlayers';
 import {TestPlayer} from '../TestPlayer';
 import {Reds} from '../../src/turmoil/parties/Reds';
+import {Greens} from '../../src/turmoil/parties/Greens';
 import {ReleaseOfInertGases} from '../../src/cards/base/ReleaseOfInertGases';
 import {JovianEmbassy} from '../../src/cards/promo/JovianEmbassy';
 import {IceAsteroid} from '../../src/cards/base/IceAsteroid';
@@ -27,6 +28,10 @@ import {SerializedTurmoil} from '../../src/turmoil/SerializedTurmoil';
 import {PoliticalAgendas} from '../../src/turmoil/PoliticalAgendas';
 import {IParty} from '../../src/turmoil/parties/IParty';
 import {GreeneryStandardProject} from '../../src/cards/base/standardProjects/GreeneryStandardProject';
+import {ArtificialLake} from '../../src/cards/base/ArtificialLake';
+import {LavaFlows} from '../../src/cards/base/LavaFlows';
+import {StripMine} from '../../src/cards/base/StripMine';
+import {GiantSolarShade} from '../../src/cards/venusNext/GiantSolarShade';
 
 describe('Turmoil', function() {
   let player : TestPlayer; let player2 : Player; let game : Game; let turmoil: Turmoil;
@@ -213,6 +218,231 @@ describe('Turmoil', function() {
     player.playedCards.push(new EarthCatapult(), new QuantumExtractor());
     player.megaCredits = 25;
     expect(player.canPlay(nitrogenFromTitan)).is.true; // 25 + 6 - 6
+  });
+
+
+  // TODO(kberg): Move tests and code to the turmoil branch.
+  it('canPlay: reds tax applies by default when raising oxygen', function() {
+  // Strip Mine raises the oxygen level two steps.
+    const card = new StripMine();
+    const player = TestPlayers.BLUE.newPlayer();
+    const game = Game.newInstance('foobar', [player], player, TestingUtils.setCustomGameOptions());
+    const turmoil = game.turmoil!;
+    game.phase = Phase.ACTION;
+    player.setProductionForTest({energy: 2}); // Card requirement.
+
+    turmoil.rulingParty = new Greens();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.true;
+
+    turmoil.rulingParty = new Reds();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.false;
+
+    player.megaCredits = card.cost + 5;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 6;
+    expect(player.canPlay(card)).is.true;
+
+    (game as any).oxygenLevel = constants.MAX_OXYGEN_LEVEL - 1;
+    player.megaCredits = card.cost + 2;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 3;
+    expect(player.canPlay(card)).is.true;
+
+    (game as any).oxygenLevel = constants.MAX_OXYGEN_LEVEL;
+
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.true;
+  });
+
+  it('canPlay: when paying reds tax for oxygen, include the cost for the 8% temperature bump.', function() {
+  // Strip Mine raises the oxygen level two steps.
+    const card = new StripMine();
+    const player = TestPlayers.BLUE.newPlayer();
+    const game = Game.newInstance('foobar', [player], player, TestingUtils.setCustomGameOptions());
+    const turmoil = game.turmoil!;
+    game.phase = Phase.ACTION;
+    player.setProductionForTest({energy: 2}); // Card requirement.
+
+    turmoil.rulingParty = new Reds();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+
+    // Raising to 8%
+    (game as any).oxygenLevel = 7;
+
+    player.megaCredits = card.cost + 8;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 9;
+    expect(player.canPlay(card)).is.true;
+  });
+
+  it('canPlay: when paying reds tax for oxygen, include the cost for the 8% temperature bump, which triggers 0° ocean bump.', function() {
+  // Strip Mine raises the oxygen level two steps.
+    const card = new StripMine();
+    const player = TestPlayers.BLUE.newPlayer();
+    const game = Game.newInstance('foobar', [player], player, TestingUtils.setCustomGameOptions());
+    const turmoil = game.turmoil!;
+    game.phase = Phase.ACTION;
+    player.setProductionForTest({energy: 2}); // Card requirement.
+
+    turmoil.rulingParty = new Reds();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+
+    // Raising to 8%
+    (game as any).oxygenLevel = 7;
+    // Raising to 0
+    (game as any).temperature = -2;
+
+    player.megaCredits = card.cost + 11;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 12;
+    expect(player.canPlay(card)).is.true;
+  });
+
+  it('canPlay: reds tax applies by default when raising temperature', function() {
+  // LavaFlows raises the temperature two steps.
+    const card = new LavaFlows();
+    const player = TestPlayers.BLUE.newPlayer();
+    const game = Game.newInstance('foobar', [player], player, TestingUtils.setCustomGameOptions());
+    const turmoil = game.turmoil!;
+    game.phase = Phase.ACTION;
+
+    turmoil.rulingParty = new Greens();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.true;
+
+    turmoil.rulingParty = new Reds();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.false;
+
+    player.megaCredits = card.cost + 5;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 6;
+    expect(player.canPlay(card)).is.true;
+
+    // Set temperature so it only raises one step.
+    (game as any).temperature = constants.MAX_TEMPERATURE - 2;
+
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 3;
+    expect(player.canPlay(card)).is.true;
+
+    (game as any).temperature = constants.MAX_TEMPERATURE;
+
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.true;
+  });
+
+  it('canPlay: when paying reds tax for temperature, include the cost for the 0° ocean bump.', function() {
+  // LavaFlows raises the temperature two steps.
+    const card = new LavaFlows();
+    const player = TestPlayers.BLUE.newPlayer();
+    const game = Game.newInstance('foobar', [player], player, TestingUtils.setCustomGameOptions());
+    const turmoil = game.turmoil!;
+    game.phase = Phase.ACTION;
+
+    turmoil.rulingParty = new Reds();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+
+    // Raising to 0
+    (game as any).temperature = -2;
+
+    player.megaCredits = card.cost + 8;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 9;
+    expect(player.canPlay(card)).is.true;
+  });
+
+  it('canPlay: reds tax applies by default when placing oceans', function() {
+  // ArtificialLake uses trSource.
+    const card = new ArtificialLake();
+    const player = TestPlayers.BLUE.newPlayer();
+    const game = Game.newInstance('foobar', [player], player, TestingUtils.setCustomGameOptions());
+    const turmoil = game.turmoil!;
+    (game as any).temperature = -6; // minimum requirement for the card.
+    game.phase = Phase.ACTION;
+
+    turmoil.rulingParty = new Greens();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.true;
+
+    turmoil.rulingParty = new Reds();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.false;
+
+    player.megaCredits = card.cost + 3;
+    expect(player.canPlay(card)).is.true;
+
+    TestingUtils.maxOutOceans(player);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.true;
+  });
+
+  // TODO(kberg): Use Towing a Comet as an example of a multi-TR thing.
+
+  it('canPlay: reds tax applies by default when raising the venus scale.', function() {
+  // GiantSolarShade raises venus three steps.
+    const card = new GiantSolarShade();
+    const player = TestPlayers.BLUE.newPlayer();
+    const game = Game.newInstance('foobar', [player], player, TestingUtils.setCustomGameOptions());
+    const turmoil = game.turmoil!;
+    game.phase = Phase.ACTION;
+
+    turmoil.rulingParty = new Greens();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.true;
+
+    turmoil.rulingParty = new Reds();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.false;
+
+    player.megaCredits = card.cost + 8;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 9;
+    expect(player.canPlay(card)).is.true;
+
+    // Set Venus so it only raises one step.
+    (game as any).venusScaleLevel = constants.MAX_VENUS_SCALE - 2;
+
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 3;
+    expect(player.canPlay(card)).is.true;
+
+    (game as any).venusScaleLevel = constants.MAX_VENUS_SCALE;
+
+    player.megaCredits = card.cost;
+    expect(player.canPlay(card)).is.true;
+  });
+
+  it('canPlay: when paying reds tax for venus, include the cost for the 16% TR', function() {
+  // GiantSolarShade raises venus three steps.
+    const card = new GiantSolarShade();
+    const player = TestPlayers.BLUE.newPlayer();
+    const game = Game.newInstance('foobar', [player], player, TestingUtils.setCustomGameOptions());
+    const turmoil = game.turmoil!;
+    game.phase = Phase.ACTION;
+
+    turmoil.rulingParty = new Reds();
+    PoliticalAgendas.setNextAgenda(turmoil, game);
+
+    // Raising to or above 16%
+    (game as any).venusScaleLevel = 14;
+
+    player.megaCredits = card.cost + 11;
+    expect(player.canPlay(card)).is.false;
+    player.megaCredits = card.cost + 12;
+    expect(player.canPlay(card)).is.true;
   });
 
   it('serializes and deserializes keeping players', function() {
